myproject/requirements.txt
asgiref==3.8.1
gunicorn==21.2.0
Django==5.1.4
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
djangorestframework==3.15.2
